Add TaskCommandFactory + refactor DI layer
