+++ src/Console/Command/SymfonyTaskCommand.php
@@ -4,6 +4,7 @@ namespace Castor\Console\Command;
 
 use Castor\Attribute\AsSymfonyTask;
 use Castor\Console\Input\GetRawTokenTrait;
+use Castor\Descriptor\SymfonyTaskDescriptor;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -29,12 +30,21 @@ class SymfonyTaskCommand extends Command
     /**
      * @param mixed[] $definition
      */
-    public function __construct(
+    private function __construct(
         public readonly AsSymfonyTask $taskAttribute,
         public readonly \ReflectionClass $class,
         public readonly array $definition,
     ) {
         parent::__construct($taskAttribute->name);
+    }
+
+    public static function createFromDescriptor(SymfonyTaskDescriptor $symfonyTaskDescriptor): self
+    {
+        return new self(
+            $symfonyTaskDescriptor->taskAttribute,
+            $symfonyTaskDescriptor->function,
+            $symfonyTaskDescriptor->definition,
+        );
     }
 
     protected function configure(): void
