+++ tests/AutocompleteTest.php
@@ -7,13 +7,12 @@ use Castor\Attribute\AsTask;
 use Castor\Console\Command\TaskCommand;
 use Castor\ContextRegistry;
 use Castor\Descriptor\TaskDescriptor;
-use Castor\EventDispatcher;
 use Castor\ExpressionLanguage;
 use Castor\Helper\Slugger;
 use Symfony\Component\Console\Completion\CompletionInput;
 use Symfony\Component\Console\Tester\CommandCompletionTester;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\String\Slugger\AsciiSlugger;
-use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 class AutocompleteTest extends TaskTestCase
 {
